Centralized config in Init, removed Hook's initialize, fixed static method calls.
