+++ src/main/java/DevSecBox/Hook.java
@@ -1,6 +1,5 @@
 package DevSecBox;
 
-import burp.api.montoya.BurpExtension;
 import burp.api.montoya.core.ByteArray;
 import burp.api.montoya.http.handler.HttpHandler;
 import burp.api.montoya.http.handler.HttpRequestToBeSent;
@@ -9,53 +8,25 @@ import burp.api.montoya.http.handler.RequestToBeSentAction;
 import burp.api.montoya.http.handler.ResponseReceivedAction;
 import burp.api.montoya.http.message.HttpHeader;
 import burp.api.montoya.http.message.responses.HttpResponse;
-import burp.api.montoya.MontoyaApi;
-import java.io.InputStream;
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ExecutorService;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
-public class Hook implements BurpExtension, HttpHandler {
+public class Hook implements HttpHandler {
     private final Map<Integer, Object[]> requestMap = new ConcurrentHashMap<>();
     private int currentRequestId = 1;
     private static final int MAX_REQUEST_ID = Integer.MAX_VALUE - 1;
-    private static ExecutorService executorService = Executors.newFixedThreadPool(1);;
-    private Map<String, String> replacements = Init.Core.WorkflowPanel.getReplacements();;
-    private List<String> nonModifiableContentTypes;
+    private static final ExecutorService executorService = Executors.newFixedThreadPool(1);
+    private Map<String, String> replacements = Init.Core.WorkflowPanel.getReplacements();
+    List<String> nonModifiableContentTypes;
     public static volatile boolean isActive = true;
-
-    @Override
-    public void initialize(MontoyaApi api) {
-        Hook.Start();
-        Init.api.logging().logToOutput(Init.PREF + Init.DSB + "orchestrator loaded - " + Linker.WORKFLOWS[0]);
-        loadConfiguration();
-        Init.api.http().registerHttpHandler(this);
-    }
-
-    private void loadConfiguration() {
-        Properties properties = new Properties();
-        try (InputStream input = getClass().getClassLoader().getResourceAsStream("config.properties")) {
-            if (input == null) {
-                throw new IOException("pre-configured types");
-            }
-            properties.load(input);
-            String types = properties.getProperty("nonModifiableContentTypes", "");
-            nonModifiableContentTypes = Arrays.asList(types.split(","));
-        } catch (IOException ex) {
-            Init.api.logging().logToOutput(Init.PREF + Init.DSB + "default settings: " + ex.getMessage());
-            nonModifiableContentTypes = List.of("image/", "application/octet-stream");
-        }
-    }
 
     @Override
     public RequestToBeSentAction handleHttpRequestToBeSent(HttpRequestToBeSent httpRequestToBeSent) {
@@ -145,17 +116,12 @@ public class Hook implements BurpExtension, HttpHandler {
             Issue.liveIssueOFF();
             Linker.processSemaphore.release();
             Linker.Scheduler.shutdownNow();
-            Init.api.logging().logToOutput(Init.PREF + Init.DSB
-                    + "hook Shutdown");
         }
     }
 
     public static void Start() {
         if (!Hook.isActive) {
             isActive = true;
-            Init.api.logging().logToOutput(Init.PREF + Init.DSB
-                    + "hook Start");
-
         }
     }
 }
