+++ src/main/java/DevSecBox/Init.java
@@ -5,12 +5,17 @@ import burp.api.montoya.extension.Extension;
 import burp.api.montoya.extension.ExtensionUnloadingHandler;
 import burp.api.montoya.logging.Logging;
 import burp.api.montoya.MontoyaApi;
+import java.io.InputStream;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Properties;
 
 public class Init implements BurpExtension {
     public static String DSB = "DevSecBox ";
     public static String PREF = "□─■ ";
     public static Logging logging;
-    private Hook Hook;
+    private Hook hook; 
     public static final OS CURRENTOS = detectOS();
     public static MontoyaApi api;
     public static Core Core;
@@ -41,8 +46,26 @@ public class Init implements BurpExtension {
         }
 
         Core = new Core();
-        Hook = new Hook();
-        Hook.initialize(api);
+        hook = new Hook();
+        Hook.Start();
+        logging.logToOutput(PREF + DSB + "orchestrator loaded and running");
+        loadConfiguration();
+        api.http().registerHttpHandler(hook);
+    }
+
+    private void loadConfiguration() {
+        Properties properties = new Properties();
+        try (InputStream input = getClass().getClassLoader().getResourceAsStream("config.properties")) {
+            if (input == null) {
+                throw new IOException("pre-configured types");
+            }
+            properties.load(input);
+            String types = properties.getProperty("nonModifiableContentTypes", "");
+            hook.nonModifiableContentTypes = Arrays.asList(types.split(","));
+        } catch (IOException ex) {
+            logging.logToOutput(PREF + DSB + "default settings: " + ex.getMessage());
+            hook.nonModifiableContentTypes = List.of("image/", "application/octet-stream");
+        }
     }
 
     private static OS detectOS() {
@@ -66,10 +89,7 @@ public class Init implements BurpExtension {
         @Override
         public void extensionUnloaded() {
             Core.WorkflowPanel.clearAllComponents();
-            logging.logToOutput(PREF + DSB + "orchestrator unloaded - LIVE HOOK ");
-            logging.logToOutput(PREF + DSB + "orchestrator unloaded - PROXY/LOGGER");
-            logging.logToOutput(PREF + DSB + "extension was unloaded.");
+            logging.logToOutput(PREF + DSB + "orchestrator stopped and unloaded");
         }
     }
-
 }
